+++ PS1_Emulator/Core/MSIL_Recompiler/CPU_MSIL_Recompiler.cs
@@ -498,10 +498,10 @@ namespace PSXEmulator {
             return ref BUS;
         }
 
-        public double GetSpeed() {
+        public long GetSpeed() {
             double returnValue = (CyclesDone / CYCLES_PER_SECOND) * 100;
             CyclesDone = 0;
-            return Math.Round(returnValue, 2);
+            return (long)Math.Round(returnValue, 2);
         }
     }
 }
